delete ludmila_v1, create ludmila
