+++ keyboards/mexsistor/ludmila/ludmila.h
@@ -27,9 +27,10 @@
  * represents the switch matrix.
  */
 #define LAYOUT( \
-    k00, k01, k02, \
-      k10,  k12    \
+    k00, k01, k20, \
+    k10, k11 \
 ) { \
-    { k00, k01,   k02 }, \
-    { k10, KC_NO, k12 }  \
-}
+    { k00, k01, }, \
+    { k10, k11, }, \
+    { k20 }  \
+  }
