+++ keyboards/mexsistor/ludmila/rules.mk
@@ -20,3 +20,6 @@ BACKLIGHT_ENABLE = no       # Enable keyboard backlight functionality
 RGBLIGHT_ENABLE = no        # Enable keyboard RGB underglow
 BLUETOOTH_ENABLE = no       # Enable Bluetooth
 AUDIO_ENABLE = no           # Audio output
+ENCODER_ENABLE = yes
+SRC = matrix.c
+CUSTOM_MATRIX = yes
