--- keyboards/mexsistor/ludmila_v1/info.json
@@ -1,19 +1,0 @@
-{
-    "keyboard_name": "ludmila_v1",
-    "url": "",
-    "maintainer": "Kevin M.",
-    "width": 3,
-    "height": 2,
-    "layouts": {
-        "LAYOUT": {
-            "layout": [
-                {"label": "k00", "x": 0, "y": 0},
-                {"label": "k01", "x": 1, "y": 0},
-                {"label": "k02", "x": 2, "y": 0},
-
-                {"label": "k10", "x": 0, "y": 1, "w": 1.5},
-                {"label": "k12", "x": 1.5, "y": 1, "w": 1.5}
-            ]
-        }
-    }
-}
