--- keyboards/mexsistor/ludmila_v1/keymaps/default/readme.md
@@ -1,1 +1,0 @@
-# The default keymap for ludmila_v1
